✨ Refactor future outfit planner UI into modular components
This commit refactors the `FutureOutfitPlannerSheet` by extracting its internal private widgets into standalone component files. This improves code maintainability and follows a more modular architecture.

- **`lib/search/ui/widgets/`**:
    - Created `header.dart`: Extracted the `_Header` widget.
    - Created `planner_form.dart`: Extracted the `_PlannerForm` widget.
    - Created `planner_result.dart`: Extracted the `_PlannerResult` widget.
    - **`future_outfit_planner_sheet.dart`**: Updated to use the new modular components and removed redundant inline widget definitions.

- **Models & Enums**:
    - Created `lib/entities/enums/temp_option.dart`: Moved the `_TempOption` enum to a public domain file for shared use across the new components.
