+++ lib/entities/enums/temp_option.dart
@@ -1,0 +1,1 @@
+enum TempOption { coolest, typical, warmest }
